# Optimize serialization reading, create benchmarks