--- a/Robust.Benchmarks/Program.cs
+++ b/Robust.Benchmarks/Program.cs
@@ -6,13 +6,6 @@ namespace Robust.Benchmarks
     {
         public static void Main(string[] args)
         {
-            // var b = new SerializationReadBenchmark();
-            //
-            // for (var i = 0; i < 100000; i++)
-            // {
-            //     b.ReadSeedDataDefinition();
-            // }
-
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run();
         }
     }