--- a/Robust.Benchmarks/Serialization/DataDefinitionWithString.cs
+++ b/Robust.Benchmarks/Serialization/DataDefinitionWithString.cs
@@ -1,0 +1,11 @@
+﻿using Robust.Shared.Serialization.Manager.Attributes;
+
+namespace Robust.Benchmarks.Serialization
+{
+    [DataDefinition]
+    public class DataDefinitionWithString
+    {
+        [field: DataField("string")]
+        private string StringField { get; } = default!;
+    }
+}