--- a/Robust.Benchmarks/Serialization/DataDefinitionWithString.cs
+++ b/Robust.Benchmarks/Serialization/DataDefinitionWithString.cs
@@ -3,7 +3,7 @@
 namespace Robust.Benchmarks.Serialization
 {
     [DataDefinition]
-    internal class DataDefinitionWithString
+    public class DataDefinitionWithString
     {
         [field: DataField("string")]
         private string StringField { get; } = default!;