--- a/Robust.Benchmarks/Serialization/SerializationReadBenchmark.cs
+++ b/Robust.Benchmarks/Serialization/SerializationReadBenchmark.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using BenchmarkDotNet.Attributes;
 using Robust.Server;
 using Robust.Shared.Configuration;
@@ -7,10 +8,11 @@ using Robust.Shared.IoC;
 using Robust.Shared.Reflection;
 using Robust.Shared.Serialization.Manager;
 using Robust.Shared.Serialization.Markdown;
+using YamlDotNet.RepresentationModel;
 
 namespace Robust.Benchmarks.Serialization
 {
-    public class SerializationReadBenchmark
+    internal class SerializationReadBenchmark
     {
         private const int N = 1000;
 
@@ -39,6 +41,11 @@ namespace Robust.Benchmarks.Serialization
 
             StringDataDefNode = new MappingDataNode();
             StringDataDefNode.AddNode(new ValueDataNode("string"), new ValueDataNode("ABC"));
+
+            var yamlStream = new YamlStream();
+            yamlStream.Load(new StringReader(SeedDataDefinition.Prototype));
+
+            SeedNode = yamlStream.Documents[0].RootNode.ToDataNodeCast<MappingDataNode>();
         }
 
         private ISerializationManager SerializationManager { get; }
@@ -48,6 +55,8 @@ namespace Robust.Benchmarks.Serialization
         private ValueDataNode IntNode { get; } = new("1");
 
         private MappingDataNode StringDataDefNode { get; }
+
+        private MappingDataNode SeedNode { get; }
 
         // [Benchmark]
         public void ReadString()
@@ -61,10 +70,16 @@ namespace Robust.Benchmarks.Serialization
             SerializationManager.ReadValue<int>(IntNode);
         }
 
-        [Benchmark]
-        public void ReadDataDefinition()
+        // [Benchmark]
+        public void ReadDataDefinitionWithString()
         {
             SerializationManager.ReadValue<DataDefinitionWithString>(StringDataDefNode);
+        }
+
+        [Benchmark]
+        public void ReadSeedDataDefinition()
+        {
+            SerializationManager.ReadValue<SeedDataDefinition>(SeedNode);
         }
     }
 }