--- a/Robust.Client/Properties/AssemblyInfo.cs
+++ b/Robust.Client/Properties/AssemblyInfo.cs
@@ -3,6 +3,7 @@
 [assembly: InternalsVisibleTo("Robust.UnitTesting")]
 [assembly: InternalsVisibleTo("Robust.Lite")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
+[assembly: InternalsVisibleTo("Robust.Benchmarks")]
 
 #if NET5_0
 [module: SkipLocalsInit]