--- a/Robust.Server/Properties/AssemblyInfo.cs
+++ b/Robust.Server/Properties/AssemblyInfo.cs
@@ -2,6 +2,7 @@
 
 [assembly: InternalsVisibleTo("Robust.UnitTesting")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
+[assembly: InternalsVisibleTo("Robust.Benchmarks")]
 
 #if NET5_0
 [module: SkipLocalsInit]