--- a/Robust.Shared/Properties/AssemblyInfo.cs
+++ b/Robust.Shared/Properties/AssemblyInfo.cs
@@ -10,6 +10,7 @@
 [assembly: InternalsVisibleTo("OpenToolkit.GraphicsLibraryFramework")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Gives access to Castle(Moq)
 [assembly: InternalsVisibleTo("Content.Benchmarks")]
+[assembly: InternalsVisibleTo("Robust.Benchmarks")]
 
 #if NET5_0
 [module: SkipLocalsInit]