--- a/Robust.Shared/Serialization/Manager/Result/DeserializationResult.cs
+++ b/Robust.Shared/Serialization/Manager/Result/DeserializationResult.cs
@@ -18,17 +18,6 @@ namespace Robust.Shared.Serialization.Manager.Result
             return (DeserializationResult) Activator.CreateInstance(type, value)!;
         }
 
-        public static DeserializationResult Definition(object value, DeserializedFieldEntry[] mappings)
-        {
-            //if (!IoCManager.Resolve<ISerializationManager>().HasDataDefinition(value.GetType()))
-            //    throw new ArgumentException("Provided value was not a data definition", nameof(value));
-
-            //todo validate mappings array count
-            //unless...
-            var type = typeof(DeserializedDefinition<>).MakeGenericType(value.GetType());
-            return (DeserializationResult) Activator.CreateInstance(type, value, mappings)!;
-        }
-
         public T Cast<T>() where T : DeserializationResult
         {
             if (this is T value) return value;