--- a/Robust.Shared/Serialization/Manager/Result/DeserializedValue.cs
+++ b/Robust.Shared/Serialization/Manager/Result/DeserializedValue.cs
@@ -1,5 +1,30 @@
 ﻿namespace Robust.Shared.Serialization.Manager.Result
 {
+    public class DeserializedValue : DeserializationResult
+    {
+        public DeserializedValue(object? value)
+        {
+            RawValue = value;
+        }
+
+        public override object? RawValue { get; }
+
+        public override DeserializationResult PushInheritanceFrom(DeserializationResult source)
+        {
+            return source.Copy().Cast<DeserializedValue>();
+        }
+
+        public override DeserializationResult Copy()
+        {
+            return new DeserializedValue(RawValue);
+        }
+
+        public override void CallAfterDeserializationHook()
+        {
+            if (RawValue is ISerializationHooks hooks) hooks.AfterDeserialization();
+        }
+    }
+
     public class DeserializedValue<T> : DeserializationResult<T>
     {
         public DeserializedValue(T value)
@@ -23,7 +48,7 @@
 
         public override void CallAfterDeserializationHook()
         {
-            if(Value is ISerializationHooks hooks) hooks.AfterDeserialization();
+            if (Value is ISerializationHooks hooks) hooks.AfterDeserialization();
         }
     }
 }