… Robust.Shared/Serialization/Manager/SerializationManager.TypeCopiers.cs
… Robust.Shared/Serialization/Manager/SerializationManager.TypeReaders.cs
… Robust.Shared/Serialization/Manager/SerializationManager.TypeSerializers.cs
… Robust.Shared/Serialization/Manager/SerializationManager.TypeValidators.cs
… Robust.Shared/Serialization/Manager/SerializationManager.TypeWriters.cs
… Robust.Shared/Serialization/Manager/SerializationManager_Typeserializers.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Custom/FlagSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/BooleanSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/ByteSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/CharSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/DecimalSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/DoubleSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/FloatSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/IntSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/LongSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/SByteSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/ShortSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/UIntSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/ULongSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Implementations/Primitive/UShortSerializer.cs
… Robust.Shared/Serialization/TypeSerializers/Interfaces/ITypeReader.cs