- Creates the Robust.Benchmarks project, as Content had one but Robust did not.
- Splits up the SerializationManager_Typeserializers partial class into TypeSerializers, TypeReaders, TypeWriters, TypeValidators, TypeCopiers and CustomSerializers partial classes.
- Uses compiled lambdas in the two places with the worst performance:
  - Resolving TryRead generic methods for the serializer lookup.
  - Creating generic data definitions.
- Creates primitive type serializers (except native int/native uint as we dont serialize those anywhere).
- Creates a non generic DeserializedValue class as creating it generically was the third worst offender and generic technology was lost in the war anyway.
- Creates a serialization read benchmark which reads a string, an int, a data definition with a single string field and a copy of the SeedPrototype data definition from Content.

Previously:
- ReadString: 3,407 ns (15x slower)
- ReadInteger: 1,283 ns (5x slower)
- ReadDataDefWithString: 11,500 ns (14x slower)
- ReadSeed: 75,600 ns (5x slower)

Now:
|                 Method |        Mean |     Error |    StdDev |
|----------------------- |------------:|----------:|----------:|
|             ReadString |    227.1 ns |   4.47 ns |   5.65 ns |
|            ReadInteger |    245.4 ns |   4.82 ns |   6.26 ns |
|  ReadDataDefWithString |    804.7 ns |  15.27 ns |  16.34 ns |
| ReadSeedDataDefinition | 15,846.8 ns | 312.89 ns | 773.39 ns |